requirements.txt
agents==1.4.0
beartype==0.9.1
carla==0.9.13
dictor==0.1.7
diskcache==5.3.0
einops==0.4.0
filterpy==1.4.5
GitPython==3.1.29
hydra-core==1.2.0
matplotlib==3.3.4
numpy==1.20.2
omegaconf==2.1.1
opencv_python==4.5.2.54
Pillow==9.2.0
py_trees==0.8.3
pyephem==9.99
pygame==2.0.1
pytorch_lightning==1.5.10
rdp==0.8
requests==2.25.1
setuptools==59.5.0
Shapely==1.7.1
simplejson==3.17.6
six==1.16.0
tabulate==0.8.9
tf==1.0.0
torch==1.9.0
torchmetrics==0.7.2
transformers==4.18.0
xmlschema==2.1.1
